pyproject.toml
[project]
name = "langchain-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.5.8",
    "langchain>=1.2.15",
    "langchain-chroma>=1.0.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.3.0",
    "langchain-google-genai>=4.2.2",
    "langchain-text-splitters>=1.1.2",
    "langgraph>=1.1.8",
    "pymupdf>=1.27.2.2",
    "pypdf>=6.10.2",
    "python-dotenv>=1.2.2",
    "sentence-transformers>=5.4.1",
    "tqdm>=4.67.1",
]
